Adding instructions on the README file
