+++ tests/test_youtube_cookies_sample_1.py
@@ -14,6 +14,7 @@ def driver():
     # This is needed so browser is not closed automatically after test
     chrome_options = Options()
     chrome_options.add_experimental_option("detach", True)
+    chrome_options.add_argument("--headless")
 
     # This is needed to specify the browser and let the browser know it won't close automatically
     driver = webdriver.Chrome(options=chrome_options)
@@ -30,16 +31,11 @@ def test_accept_button_closes_cookies_concept_popup(driver):
     driver.get(BASE_URL)
 
     # This search the element and store it into a variable, XPATHs locators are not recommended ⚠️
-   # accept_all = driver.find_element(By.XPATH,
-    #                                  '//*[@id="content"]/div[2]/div[6]/div[1]/ytd-button-renderer[2]/yt-button-shape/button/yt-touch-feedback-shape/div/div[2]')
-
-    accept_all = By.XPATH, '//*[@id="content"]/div[2]/div[6]/div[1]/ytd-button-renderer[2]/yt-button-shape/button/yt-touch-feedback-shape/div/div[2]'
-
-    WebDriverWait(driver, 10).until(EC.visibility_of_element_located(accept_all)).click()
+    accept_all = (By.XPATH, '//*[@id="content"]/div[2]/div[6]/div[1]/ytd-button-renderer[2]/yt-button-shape/button/yt-touch-feedback-shape/div/div[2]')
 
     # accept_all.click()
 
     # After the click, we want to give time to the browser to close the popup, so our assertion is not flaky
-    WebDriverWait(driver, 10).until(EC.invisibility_of_element(accept_all))
+    WebDriverWait(driver, 15).until(EC.invisibility_of_element(accept_all))
 
-    assert not driver.find_element(accept_all).is_displayed()
+    assert not driver.find_element(*accept_all).is_displayed()
